feat + fix: UnauthenticatedURLs 클래스 추가 및 패키지 구조 수정
- 테스트 및 객체의 책임 분리 위해 로그인이 필요없는 경로를 클래스로 관리하도록 하였음
