+++ src/main/java/webserver/authenticator/Authenticator.java
@@ -1,43 +1,26 @@
 package webserver.authenticator;
 
-import static webserver.requesthandler.URLConst.CSS_FILE_URL;
-import static webserver.requesthandler.URLConst.HOME_URL;
-import static webserver.requesthandler.URLConst.IMG_FILE_URL;
-import static webserver.requesthandler.URLConst.LOGIN_URL;
-import static webserver.requesthandler.URLConst.REGISTRATION_URL;
-
 import java.io.IOException;
-import java.util.HashSet;
-import java.util.Set;
 import model.User;
 import webserver.http.HttpRequest;
 import webserver.http.HttpResponse;
+import webserver.requesthandler.url.UnauthenticatedURLs;
 import webserver.session.SessionManager;
 
 public class Authenticator {
-    private final Set<String> unauthenticatedURLs; // 로그인이 필요 없는 경로 모음
+    private final UnauthenticatedURLs unauthenticatedURLs; // 로그인이 필요 없는 경로 모음
 
-    public Authenticator() {
-        this.unauthenticatedURLs = initUnauthenticatedUrls();
-    }
-
-    private Set<String> initUnauthenticatedUrls() {
-        Set<String> unauthenticatedUrls = new HashSet<>();
-        unauthenticatedUrls.add(HOME_URL);
-        unauthenticatedUrls.add(LOGIN_URL);
-        unauthenticatedUrls.add(REGISTRATION_URL);
-        unauthenticatedUrls.add(IMG_FILE_URL);
-        unauthenticatedUrls.add(CSS_FILE_URL);
-        return unauthenticatedUrls;
+    public Authenticator(UnauthenticatedURLs unAuthenticatedURLs) {
+        this.unauthenticatedURLs = unAuthenticatedURLs;
     }
 
     public boolean isAuthenticated(HttpRequest request, HttpResponse response) throws IOException {
         String requestPath = request.getPath();
 
-        if (isLoginCheckPath(requestPath)) { // 로그인이 필요한 경로인지 검사
+        if (unauthenticatedURLs.isLoginCheckPath(requestPath)) { // 로그인이 필요한 경로인지 검사
             boolean isLoggedIn = isLoggedIn(request);
             if (!isLoggedIn) { // 로그인이 안되어 있으면 로그인 화면으로 리다이렉트
-                response.sendRedirect(LOGIN_URL + "?redirectURL=" + requestPath);
+                response.sendRedirect("/login?redirectURL=" + requestPath);
                 return false;
             }
         }
@@ -49,15 +32,5 @@ public class Authenticator {
         // 로그인 상태 확인
         User user = (User) SessionManager.findSession(request);
         return user != null;
-    }
-
-    private boolean isLoginCheckPath(String requestPath) {
-        // 로그인이 필요한 경로인지 확인
-        return unauthenticatedURLs.stream()
-                .noneMatch(mappedURL -> isMatched(mappedURL, requestPath));
-    }
-
-    private boolean isMatched(String mappedURL, String requestPath) {
-        return requestPath.matches(mappedURL);
     }
 }
