+++ src/main/java/webserver/requesthandler/MainRequestHandler.java
@@ -1,12 +1,12 @@
 package webserver.requesthandler;
 
-import static webserver.requesthandler.URLConst.ARTICLE_URL;
-import static webserver.requesthandler.URLConst.COMMENT_URL;
-import static webserver.requesthandler.URLConst.HOME_URL;
-import static webserver.requesthandler.URLConst.LOGIN_URL;
-import static webserver.requesthandler.URLConst.LOGOUT_URL;
-import static webserver.requesthandler.URLConst.REGISTRATION_URL;
-import static webserver.requesthandler.URLConst.USER_LIST_URL;
+import static webserver.requesthandler.url.URLConst.ARTICLE_URL;
+import static webserver.requesthandler.url.URLConst.COMMENT_URL;
+import static webserver.requesthandler.url.URLConst.HOME_URL;
+import static webserver.requesthandler.url.URLConst.LOGIN_URL;
+import static webserver.requesthandler.url.URLConst.LOGOUT_URL;
+import static webserver.requesthandler.url.URLConst.REGISTRATION_URL;
+import static webserver.requesthandler.url.URLConst.USER_LIST_URL;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -19,6 +19,16 @@ import org.slf4j.LoggerFactory;
 import webserver.authenticator.Authenticator;
 import webserver.http.HttpRequest;
 import webserver.http.HttpResponse;
+import webserver.requesthandler.handlerimpl.ArticleHandler;
+import webserver.requesthandler.handlerimpl.CommentHandler;
+import webserver.requesthandler.handlerimpl.HomeHandler;
+import webserver.requesthandler.handlerimpl.LoginHandler;
+import webserver.requesthandler.handlerimpl.LogoutHandler;
+import webserver.requesthandler.handlerimpl.RegistrationHandler;
+import webserver.requesthandler.handlerimpl.RequestHandler;
+import webserver.requesthandler.handlerimpl.StaticResourceHandler;
+import webserver.requesthandler.handlerimpl.UserListHandler;
+import webserver.requesthandler.url.UnauthenticatedURLs;
 
 public class MainRequestHandler implements Runnable {
     private static final Logger logger = LoggerFactory.getLogger(MainRequestHandler.class);
@@ -40,7 +50,7 @@ public class MainRequestHandler implements Runnable {
             HttpRequest request = new HttpRequest(in);
             HttpResponse response = new HttpResponse(out);
 
-            Authenticator authenticator = new Authenticator();
+            Authenticator authenticator = new Authenticator(new UnauthenticatedURLs());
             boolean isAuthenticated = authenticator.isAuthenticated(request, response);// 인증이 필요한 페이지에 접근하는지 확인
 
             if (!isAuthenticated) {
