+++ src/main/java/webserver/requesthandler/handlerimpl/CommentHandler.java
@@ -1,13 +1,14 @@
-package webserver.requesthandler;
+package webserver.requesthandler.handlerimpl;
 
 import java.io.IOException;
 import webserver.http.HttpRequest;
 import webserver.http.HttpResponse;
+import webserver.requesthandler.url.URLConst;
 
-public class ArticleHandler implements RequestHandler {
+public class CommentHandler implements RequestHandler {
     @Override
     public void handleGet(HttpRequest request, HttpResponse response) throws IOException {
-        setHTMLToBody(response, URLConst.ARTICLE_URL);
+        setHTMLToBody(response, URLConst.COMMENT_URL);
     }
 
     @Override
