+++ src/main/java/webserver/requesthandler/handlerimpl/StaticResourceHandler.java
@@ -1,4 +1,4 @@
-package webserver.requesthandler;
+package webserver.requesthandler.handlerimpl;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -6,6 +6,7 @@ import utils.FileManager;
 import webserver.http.ContentType;
 import webserver.http.HttpRequest;
 import webserver.http.HttpResponse;
+import webserver.requesthandler.url.URLConst;
 
 public class StaticResourceHandler implements RequestHandler {
